+++ app/controllers/family_memberships_controller.rb
@@ -2,6 +2,7 @@
 
 class FamilyMembershipsController < ApplicationController
   before_action :authenticate_user!
+  before_action :ensure_family_feature_enabled!
   before_action :set_family
   before_action :set_membership, only: %i[show destroy]
 
@@ -29,6 +30,12 @@ class FamilyMembershipsController < ApplicationController
   end
 
   private
+
+  def ensure_family_feature_enabled!
+    unless DawarichSettings.family_feature_enabled?
+      redirect_to root_path, alert: 'Family feature is not available'
+    end
+  end
 
   def set_family
     @family = current_user.family